requirements.txt
python-telegram-bot==21.0
APScheduler==3.10.4
pytz==2023.3
python-dotenv==1.0.0 